Fix build error when SSL configuration is disabled
When QT_NO_SSL is used, the value of QT_FEATURE_ssl is also defined to
'-1' in qtnetwork-config.h. The "#if defined(QT_FEATURE_ssl)" block will
be compiled, and then it implies that 'QSslSocket' was not declared in
this scope.

The QT_CONFIG macro implements a compile-time check for features of Qt.
If QT_FEATURE_ssl is '0' or undefined, QT_ CONFIG(ssl) will lead to a
compile error. If QT_FEATURE_ssl is '-1' (unavailable), the
constant-expression of QT_ CONFIG(ssl) will be false.

Change-Id: I80bef731e8246f9206601527435a43195989331b
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>
Reviewed-by: Jesus Fernandez <Jesus.Fernandez@qt.io>